Better integration of different parts
Added a KParts::PartManager to each QuadKonsole window to merge the
KPart's menus and toolbars into QuadKonsole's.

Added an autogenerated popup menu for each KPart that contains all it's
exposed actions. Dolphin's actions do not always contain group names, so
the menu looks a bit strange but works.

If a KPart does not support opening some URL based on it's mime type,
QKStack tries to find another suitable KPart in it's current stack. If
no part supports the given type, QKStack tries again with the containing
folder of the given URL.

QuadKonsole exposes the status bar and window title to the current
active KPart.