Fixed opening of non-local URLs
All mime type getting/checking is performed by QKStack. The almighty KIO
will lookup types of remote URLs (those with protocol != file) for us
and QKStack knows, if the user want's to see another KPart or to open
the requested URL in the current view. If the current view does not like
the given mime type, the URL is still passed on like before.