RMB menus, history and fixes
The RMB menus for KParts with BrowserExtension are generated by libkonq
now (with fallback to old style if libkonq is not available at build
time).

Added middle click action on QKStack's tabs to close selected tab and a
shortcut (CTRL + SHIFT + Q) to close current view.

If a KPart's factory can't be created, there will only be one error
displayed. Not one error per try.

Fixed a crash when history is empty on switching views.

Fixed closing a window if a Konsole KPart was closed before.

Fixed startup if primary KPart is not available.