Added zoom view and corrected queryClose
Fixed a crash on detaching when updating the history. Fixed a crash on
detaching when resetting the active stack and reconnecting URL bar
actions.

Corrected the queryClose functions. They check for modified KParts and
running processes.

Rewrote the detach on quit dialog to allow detaching of background
views. It also displays the part name and current URL.