Editable contents in KParts
Some of the KParts that I regularly test provide both, a
KParts::ReadOnlyPart and a KParts::ReadWritePart. The ReadWritePart is
required to change anything inside the displayed document (e.g. some
text in katepart). QKView checks for service type provided by the given
KPart and instantiates a KParts::ReadWritePart when possible. It also
deactivates editing on loading. A new action in the edit menu toggles
the read/write flag.

Added the size of the sidebar to session information and some short
aliases for commandline options.