Open URLs in suitable application if there is no KPart
If KMimeTypeTrader does not find a KPart to open a given URL, QKStack will try to run a suitable application for it. There is no config option for this behavior and might lead to undesirable results if Settings::autoloadView() is not set.

Small fix for session restoration (copied from Konsole).

Defined initial size for the sidebar splitter if it can not be restored from auto saved config.