Fixed URL handling if given on command line
Keep given index if prepending working directory. Support for umlauts in URLs (encoding set to URF-8).

Select the main window with input focus or the newest one for opening URLs from outside of QuadKonsole4.

If no index is given in URL to open, use the zoomed stack, the active (focused) stack or the next available stack (tried in that order).