Added config option for default Konsole profile
This is a dirty hack! Konsolepart does not provide an API to load a
named profile on startup, change to a named profile or query the list of
defined profiles. This hack searches for all *.profile files in
Konsole's data directories and generates a single config string for
changeSessionSettings(). Works for Konsole-2.6.4 and maybe other
versions.