Fixed DBUS name in quadkonsole4.desktop
Without defined X-DBUS-ServiceName in the desktop file, kdeinit did not find the running instance of QuadKonsole4 after starting another process. The first start worked, but adding another window lead to the message "kdeinit could not launch...".

Fixed duplicating of current view when opening a URL from outside. As long a the caller is not waiting for the new widget, the QKStack resolves the MIME type before creating a new QKView.